refactor(orchestrator): split monolithic module + extract stats/test helpers
Address all 7 architectural findings from age review on PR #66, plus
S3 (BUG) on _extract_abandon_reason direction:

- A1/A2/A4: split orchestrator.py (368L) into a package with focused
  submodules (_phase, _abandon, _tune, _generate). orchestrate_tune and
  orchestrate_generate now compose small helpers under the line + param
  budgets; the package facade re-exports public + selected private
  symbols so existing imports are preserved.
- A3: extract _query_filtered helper in analytics to dedupe the three
  cwd-filter branches in _build_patterns_text.
- A5: move stats rendering out of __main__.py into _stats_rendering.py
  so the CLI dispatcher stays under budget.
- A6: split tests/test_orchestrator.py (1217L) into focused suites:
  test_parse_judge_signal, test_abandon_rule, test_orchestrate_tune,
  test_orchestrate_generate, test_cli_rewire. Shared FakeRalphRunner
  now lives in conftest.py.
- S3 (BUG): _extract_abandon_reason searches bottom-up so the prose
  above the *final* JUDGE_* signal is returned, matching the docstring
  and JudgeParseError's bottom-up parsing.

A7 (_run_phase 5 params) deferred — reducing it would force a bundling
type that hurts readability; existing test asserts the 5-positional-arg
signature.

just build green: 942 tests passing, 99% diff coverage.
