+++ vaudeville/__main__.py
@@ -12,13 +12,12 @@ import sys
 from typing import Any
 
 from rich.console import Console
-from rich.text import Text
 
 from vaudeville import orchestrator
+from vaudeville._stats_rendering import print_stats_human
 from vaudeville.cli_rules import attach_rule_parsers, dispatch_rule_command
 from vaudeville.core.paths import find_project_root as _core_find_project_root
 from vaudeville.orchestrator import RalphError, Thresholds
-from vaudeville.server import latency_text, styled_table
 
 
 _EVENTS_LOG = os.path.join(
@@ -136,71 +135,8 @@ def cmd_generate(args: argparse.Namespace) -> int:
         return 2
 
 
-def _render_rules_table(con: Console, rules: dict[str, Any], total: int) -> None:
-    table = styled_table(
-        "Per-Rule Breakdown",
-        caption=f"{total} classifications",
-    )
-    table.add_column("Rule", justify="left")
-    for col in ("Total", "Violations", "Pass %", "Avg ms", "p50 ms", "p95 ms"):
-        table.add_column(col, justify="right")
-
-    for name, data in rules.items():
-        pass_rate = data["pass_rate"]
-        style = "green" if pass_rate >= 90 else "yellow" if pass_rate >= 50 else "red"
-        table.add_row(
-            name,
-            str(data["total"]),
-            str(data["violations"]),
-            Text(f"{pass_rate:.1f}%", style=style),
-            latency_text(data["avg_latency_ms"]),
-            latency_text(data["p50_latency_ms"]),
-            latency_text(data["p95_latency_ms"]),
-        )
-    con.print(table)
-
-
-def _render_latency_summary(con: Console, lat: dict[str, Any]) -> None:
-    summary = Text.assemble(
-        "Overall latency \u2014 p50: ",
-        (f"{lat['p50_ms']:.1f}ms", "bold"),
-        "   p95: ",
-        (f"{lat['p95_ms']:.1f}ms", "bold"),
-        "   mean: ",
-        (f"{lat['mean_ms']:.1f}ms", "bold"),
-    )
-    con.print(summary)
-    con.print()
-
-
-def _render_histogram(con: Console, histogram: dict[str, int]) -> None:
-    hist_table = styled_table("Latency Histogram")
-    hist_table.add_column("Bucket", justify="right")
-    hist_table.add_column("Count", justify="right")
-    hist_table.add_column("Bar")
-
-    max_count = max(histogram.values()) if histogram else 1
-    bar_width = 40
-    for bucket, count in histogram.items():
-        bar_len = int((count / max_count) * bar_width) if max_count else 0
-        hist_table.add_row(bucket, str(count), "\u2588" * bar_len)
-    con.print(hist_table)
-
-
 def _print_stats_human(result: dict[str, Any], console: Console | None = None) -> None:
-    con = console if console is not None else _console
-    total = result["total"]
-    if total == 0:
-        con.print("No events recorded yet.")
-        return
-
-    con.rule("Vaudeville Stats", style="bold cyan")
-    rules = result["rules"]
-    if rules:
-        _render_rules_table(con, rules, total)
-    lat = result["latency"]
-    _render_latency_summary(con, lat)
-    _render_histogram(con, lat["histogram"])
+    print_stats_human(result, console if console is not None else _console)
 
 
 def _add_log_path_arg(p: argparse.ArgumentParser) -> None:
